Change the internal API.
We now use a single client, to fetch many documents. This makes a lot more
sense, and even allows a persistent instance of the HTTPClient, so connections
can be kept alive.

I plan on exposing this as the final API in the future, it just needs some
factory methods to build a cache object as part of the client.