Model the Document consistently.
Previously, it was either taking a string, or a HTTP response object, then
storing that, and each accessor method would branch to handle both cases.

This was madness.

The string representation is gone. As it was only used for serialisation, we
have replaced it with a YAML version.

The HTTP response version is also gone, because we don't want to hold onto data
we don't care about. Instead we have a simple class, with a factory method that
can create an instance from a HTTP response.